requirements.txt
torch==2.8.0
numpy>=2.1,<3
tqdm==4.65.0
transformers>=4.41,<4.46
tokenizers>=0.15,<0.20
lightning>=2.3,<2.5
pillow>=10.3,<12
